requirements.txt
setuptools~=58.0.4
torch~=1.10.0
torchvision~=0.11.1
numpy~=1.21.4
pillow~=8.4.0
scikit-learn~=1.0.1
imgaug~=0.4.0
pandas~=1.3.4
matplotlib~=3.5.0
jsonlines~=2.0.0
boto3~=1.21.10
tqdm~=4.62.3
croissant-ml
pytest~=6.2.5
argschema~=2.0.2
marshmallow~=3.0.0rc6
sagemaker~=2.77.1
opencv-python~=4.5.4.58
docker